OS X: Simplify the launching of Wings3D
On OS X, there would be two Wings icons while Wings was running.  The
reason is that there was a main executable (a wrapper) that started
the Erlang/OTP virtual machine with Wings, and both remained running.

It turns out that the main executable in an application can be
a script. So we can make a modified version of the 'erl' script
the main executable for the Wings3D application.

It is not obvious how to use Xcode to create an application whose main
executable is not an Objective-C program. The easiest solution is to
not use Xcode, but to modify the mac_release script to create the
Wings3D.app folder with all of its contents required by an OS X
application. Most of shell script code for copying files can be copied
from the Xcode project.

NOTE: On Mac OS X, there used two appera two Wings3D icons in
the Dock when Wings3D was launched. The extra icon has been eliminated.
[bjorng]