Wings3D.sh: Sanitize the environment before starting Wings
On an computer where Wings development had been done, the
ERL_LIBS environment variable would probably be set. If the
released version of Wings was run on such a computer, the
result could be a crash or that a different version of Wings
was run.

Other environment variables recognized by the Erlang run-time
system could potentially also cause problems. Therefore we will
unset all environment variables that start with "ERL_".